+++ gegelatilib/include/learn/learningParameters.h
@@ -36,10 +36,10 @@
 #ifndef LEARNING_PARAMETERS_H
 #define LEARNING_PARAMETERS_H
 
-#include <cstdint>
 #include <thread>
-
 #include "mutator/mutationParameters.h"
+
+
 
 namespace Learn {
 	/**